🐛 텍스트 버튼 커서 모양 적용
! 로 버튼일 경우 강제 적용
